AI_Image_PPT_Converter/requirements.txt
openai>=1.0
python-pptx>=0.6.21
pillow>=10.0.0
python-dotenv>=1.0.0
flask>=3.0.0